# api/openapi.yaml
openapi: 3.1.0
info:
  title: Ydisks 闲鱼助手 API
  version: 1.0.0
  description: 版本化 HTTP API 的唯一契约源。
servers:
  - url: /
paths:
...
  "/api/v1/chat/images":
    post:
      operationId: postApiV1ChatImages
      summary: postApiV1ChatImages
      security: [{cookieAuth: []}]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [account_id, chat_id, buyer_id, image]
              properties:
                account_id: {type: string}
                chat_id: {type: string}
                buyer_id: {type: string}
                buyer_name: {type: string}
                buyer_avatar_url: {type: string}
                item_id: {type: string}
                item_title: {type: string}
                image: {type: string, format: binary}
      responses:
        '201':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatMessageEnvelope'
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  "/api/v1/chat/messages":
    get:
      operationId: getApiV1ChatMessages
      summary: getApiV1ChatMessages
      security: [{cookieAuth: []}]
      parameters:
        - name: account_id
          in: query
          required: true
          schema: {type: string}
        - name: chat_id
          in: query
          required: true
          schema: {type: string}
        - name: cursor
          in: query
          required: false
          schema: {type: integer}
        - name: before_id
          in: query
          required: false
          schema: {type: integer}
        - name: limit
          in: query
          required: false
          schema: {type: integer, minimum: 1, maximum: 200}
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatMessagePage'
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: postApiV1ChatMessages
      summary: postApiV1ChatMessages
      security: [{cookieAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatMessageRequest'
      responses:
        '201':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatMessageEnvelope'
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  "/api/v1/chat/read":
    post:
      operationId: postApiV1ChatRead
      summary: postApiV1ChatRead
      security: [{cookieAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatReadRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  "/api/v1/chat/sessions":
    get:
      operationId: getApiV1ChatSessions
      summary: getApiV1ChatSessions
      security: [{cookieAuth: []}]
      parameters:
        - name: account_id
          in: query
          required: true
          schema: {type: string}
        - name: cursor
          in: query
          required: false
          schema: {type: integer}
        - name: refresh
          in: query
          required: false
          schema: {type: integer, enum: [0, 1]}
        - name: limit
          in: query
          required: false
          schema: {type: integer, minimum: 1, maximum: 200}
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatSessionPage'
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  "/api/v1/chat/ws":
    get:
      operationId: getApiV1ChatWs
      summary: getApiV1ChatWs
      security: [{cookieAuth: []}]
      x-contract-success-exception:
        kind: websocket
      x-websocket-message-schema:
        $ref: '#/components/schemas/ChatWebSocketEvent'
      responses:
        '101':
          description: WebSocket 握手成功
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
...
    ChatSession:
      type: object
      required: [account_id, chat_id, buyer_id, buyer_name, last_message, last_message_at, unread_count]
      properties:
        account_id: {type: string}
        chat_id: {type: string}
        buyer_id: {type: string}
        buyer_name: {type: string}
        buyer_avatar_url: {type: string}
        item_id: {type: string}
        item_title: {type: string}
        last_message: {type: string}
        last_message_at: {type: integer}
        unread_count: {type: integer}
    ChatMessage:
      type: object
      required: [id, account_id, chat_id, message_key, direction, sender_id, sender_name, message_type, content, status, sent_at]
      properties:
        id: {type: integer}
        account_id: {type: string}
        chat_id: {type: string}
        message_key: {type: string}
        direction: {type: string, enum: [incoming, outgoing]}
        sender_id: {type: string}
        sender_name: {type: string}
        message_type: {type: string, enum: [text, image, video, system]}
        content: {type: string}
        status: {type: string, enum: [received, sending, sent, failed]}
        read_status: {type: integer}
        read_at: {type: integer}
        sent_at: {type: integer}
    ChatSessionPage:
      type: object
      required: [sessions, has_more]
      properties:
        sessions: {type: array, items: {$ref: '#/components/schemas/ChatSession'}}
        has_more: {type: boolean}
        next_cursor: {type: integer}
    ChatMessagePage:
      type: object
      required: [messages, has_more]
      properties:
        messages: {type: array, items: {$ref: '#/components/schemas/ChatMessage'}}
        has_more: {type: boolean}
        next_cursor: {type: integer}
        session: {$ref: '#/components/schemas/ChatSession'}
    ChatMessageEnvelope:
      type: object
      required: [message]
      properties:
        message: {$ref: '#/components/schemas/ChatMessage'}
    ChatReadyEvent:
      type: object
      required: [type, at]
      properties:
        type: {type: string, enum: [ready]}
        at: {type: integer}
    ChatEvent:
      type: object
      required: [type]
      properties:
        type: {type: string, enum: [message.created, message.updated]}
        message: {$ref: '#/components/schemas/ChatMessage'}
        session: {$ref: '#/components/schemas/ChatSession'}
    ChatWebSocketEvent:
      anyOf:
        - $ref: '#/components/schemas/ChatReadyEvent'
        - $ref: '#/components/schemas/ChatEvent'
    ChatMessageRequest:
      type: object
      required: [account_id, chat_id, buyer_id, text]
      properties:
        account_id: {type: string}
        chat_id: {type: string}
        buyer_id: {type: string}
        buyer_name: {type: string}
        item_id: {type: string}
        item_title: {type: string}
        text: {type: string, maxLength: 2000}
    ChatReadRequest:
      type: object
      required: [account_id, chat_id]
      properties:
        account_id: {type: string}
        chat_id: {type: string}
        message_ids: {type: array, items: {type: object, additionalProperties: {oneOf: [{type: string}, {type: number}]}}}
...
    OperationResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: string
        requires_relogin:
          type: boolean
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        request_id:
          type: string
        details:
          type: object
      additionalProperties: true